requirements.txt
aws-lambda-powertools[all]
python-binance
boto3
pydantic